refactor: update new logo
