+++ src/MainWindow/HomePage/AboutModal.jsx
@@ -4,7 +4,7 @@ import arrowRepeat from '../../assets/arrow-repeat-white.svg'
 import boxRightArrow from '../../assets/box-arrow-up-right-purple.svg'
 import labModeOff from '../../assets/lab-icon-off.svg'
 import labModeOn from '../../assets/lab-icon-on.svg'
-import phLogo from '../../assets/oobee-logo.svg'
+import phLogo from '../../assets/logo-oobee-full-colour-inverse.svg'
 import { handleClickLink, versionComparator } from '../../common/constants'
 import { useEffect, useState } from 'react'
 import Button from '../../common/components/Button'
